requirements.txt
dash
pandas
gunicorn
sqlalchemy
pymysql
python-dotenv
dash_bootstrap_components
